pyproject.toml
[project]
name = "mixseek-plus"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
authors = [
    { name = "driller", email = "eleshis@gmail.com" }
]
requires-python = ">=3.13.7"
dependencies = [
    "mixseek-core",
    "pydantic-ai",
    "tavily-python>=0.7.4",
]

[project.scripts]
mixseek = "mixseek_plus.cli:main"
mixseek-plus = "mixseek_plus:main"

[build-system]
requires = ["uv_build>=0.9.24,<0.10.0"]
build-backend = "uv_build"

[dependency-groups]
dev = [
    "mypy>=1.19.1",
    "pytest>=9.0.2",
    "pytest-asyncio>=1.0.0",
    "ruff>=0.14.13",
]
docs = [
    "myst-parser>=5.0.0",
    "sphinx>=9.1.0",
    "sphinx-rtd-theme>=3.1.0",
    "sphinxcontrib-mermaid>=2.0.0",
]

[tool.uv.sources]
mixseek-core = { git = "https://github.com/mixseek/mixseek-core.git" }
